Move shortcake UI plugin integration into the main file
